implemented optimization suggested by ecdr in issue 489 updated eusci implementation from hard coded spi port to use the definition of the pins_energia.h file instead corrected spi pin definitions for FR5969
